Add files via upload
Playing around with the automobile data set from UCI ML to see the effects of horsepower and body style on prices of the car. 

Are you actually paying premium just to get the tag of "sports car"? Can you get the same bang for the buck if you let go body-style?